add a lot of docs
