+++ src/common/src/util/tracing.rs
@@ -20,6 +20,22 @@ use tracing_opentelemetry::OpenTelemetrySpanExt;
 
 /// Context for tracing used for propagating tracing information in a distributed system.
 ///
+/// Generally, the caller of a service should create a tracing context from the current tracing span
+/// and pass it to the callee through the network. The callee will then attach its local tracing
+/// span as a child of the tracing context, so that the external tracing service can associate them
+/// in a single trace.
+///
+/// The tracing context must be serialized and deserialized when passing through the network.
+/// There're two ways to do this:
+///
+/// - For RPC calls with clear caller/callee relationship, the tracing context can be serialized
+///   into the W3C trace context format and passed in HTTP headers simply with a service middleware.
+///   For example, the DDL requests from the frontend to the meta service.
+/// - For RPC calls with no clear caller/callee relationship or with asynchronous semantics, the
+///   tracing context should be passed manually as a protobuf field in the request and response for
+///   better lifetime management. For example, the exchange services between streaming actors or
+///   batch stages, and the `create_task` requests from the frontend to the batch service.
+///
 /// See [Trace Context](https://www.w3.org/TR/trace-context/) for more information.
 #[derive(Debug, Clone)]
 pub struct TracingContext(opentelemetry::Context);
@@ -48,36 +64,47 @@ impl TracingContext {
         span
     }
 
+    /// Convert the tracing context to the W3C trace context format.
     fn to_w3c(&self) -> HashMap<String, String> {
         let mut fields = HashMap::new();
         Propagator::new().inject_context(&self.0, &mut fields);
         fields
     }
 
+    /// Create a new tracing context from the W3C trace context format.
     fn from_w3c(fields: &HashMap<String, String>) -> Self {
         let context = Propagator::new().extract(fields);
         Self(context)
     }
 
+    /// Convert the tracing context to the protobuf format.
     pub fn to_protobuf(&self) -> HashMap<String, String> {
         self.to_w3c()
     }
 
+    /// Create a new tracing context from the protobuf format.
     pub fn from_protobuf(fields: &HashMap<String, String>) -> Self {
         Self::from_w3c(fields)
     }
 
+    /// Convert the tracing context to the W3C trace context format in HTTP headers.
     pub fn to_http_headers(&self) -> http::HeaderMap {
         let map = self.to_w3c();
         http::HeaderMap::try_from(&map).unwrap_or_default()
     }
 
+    /// Create a new tracing context from the W3C trace context format in HTTP headers.
+    ///
+    /// Returns `None` if the headers are invalid.
     pub fn from_http_headers(headers: &http::HeaderMap) -> Option<Self> {
         let mut map = HashMap::new();
+
+        // See [Trace Context](https://www.w3.org/TR/trace-context/) for these header names.
         for key in ["traceparent", "tracestate"] {
             let value = headers.get(key)?.to_str().ok()?;
             map.insert(key.to_string(), value.to_string());
         }
+
         Some(Self::from_w3c(&map))
     }
 }
